Changed tests and methods in craps to doubles

+++ src/test/java/com/github/zipcodewilmington/GameTests/CrapsTests/CrapsTest.java
@@ -10,10 +10,10 @@ public class CrapsTest {
     public void testAnySeven1(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anySeven(2, 5, 100);
+        double actual = craps.anySeven(2, 5, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -23,10 +23,10 @@ public class CrapsTest {
     public void testAnySeven2(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anySeven(5, 2, 100);
+        double actual = craps.anySeven(5, 2, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -36,10 +36,10 @@ public class CrapsTest {
     public void testAnySeven3(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anySeven(4, 3, 100);
+        double actual = craps.anySeven(4, 3, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -49,10 +49,10 @@ public class CrapsTest {
     public void testAnySeven4(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anySeven(3, 4, 100);
+        double actual = craps.anySeven(3, 4, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -62,10 +62,10 @@ public class CrapsTest {
     public void testAnySeven5(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anySeven(1, 6, 100);
+        double actual = craps.anySeven(1, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -78,7 +78,7 @@ public class CrapsTest {
         int expected = 700;
 
         //When
-        int actual = craps.anySeven(6, 1, 100);
+        double actual = craps.anySeven(6, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -88,10 +88,10 @@ public class CrapsTest {
     public void testAnySevenLoss(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 0;
+        double expected = 0;
 
         //When
-        int actual = craps.anySeven(1, 1, 100);
+        double actual = craps.anySeven(1, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -101,10 +101,10 @@ public class CrapsTest {
     public void testBigSixAndEight1(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(3, 3, 100);
+        double actual = craps.bigSixAndEight(3, 3, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -114,10 +114,10 @@ public class CrapsTest {
     public void testBigSixAndEight2(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(5, 1, 100);
+        double actual = craps.bigSixAndEight(5, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -127,10 +127,10 @@ public class CrapsTest {
     public void testBigSixAndEight3(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(1, 5, 100);
+        double actual = craps.bigSixAndEight(1, 5, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -140,10 +140,10 @@ public class CrapsTest {
     public void testBigSixAndEight4(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(4, 2, 100);
+        double actual = craps.bigSixAndEight(4, 2, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -153,10 +153,10 @@ public class CrapsTest {
     public void testBigSixAndEight5(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(2, 4, 100);
+        double actual = craps.bigSixAndEight(2, 4, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -166,10 +166,10 @@ public class CrapsTest {
     public void testBigSixAndEight6(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(4, 4, 100);
+        double actual = craps.bigSixAndEight(4, 4, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -179,10 +179,10 @@ public class CrapsTest {
     public void testBigSixAndEight7(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(5, 3, 100);
+        double actual = craps.bigSixAndEight(5, 3, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -192,10 +192,10 @@ public class CrapsTest {
     public void testBigSixAndEight8(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(3, 5, 100);
+        double actual = craps.bigSixAndEight(3, 5, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -205,10 +205,10 @@ public class CrapsTest {
     public void testBigSixAndEight9(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(6, 2, 100);
+        double actual = craps.bigSixAndEight(6, 2, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -218,10 +218,10 @@ public class CrapsTest {
     public void testBigSixAndEight10(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 200;
+        double expected = 200;
 
         //When
-        int actual = craps.bigSixAndEight(2, 6, 100);
+        double actual = craps.bigSixAndEight(2, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -231,10 +231,10 @@ public class CrapsTest {
     public void testBigSixAndEightLoss(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 0;
+        double expected = 0;
 
         //When
-        int actual = craps.bigSixAndEight(1, 1, 100);
+        double actual = craps.bigSixAndEight(1, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -244,10 +244,10 @@ public class CrapsTest {
     public void testHardWays1(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 900;
+        double expected = 900;
 
         //When
-        int actual = craps.hardWays(3, 3, 100);
+        double actual = craps.hardWays(3, 3, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -257,10 +257,10 @@ public class CrapsTest {
     public void testHardWays2(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 900;
+        double expected = 900;
 
         //When
-        int actual = craps.hardWays(4, 4, 100);
+        double actual = craps.hardWays(4, 4, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -270,10 +270,10 @@ public class CrapsTest {
     public void testHardWays3(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.hardWays(5, 5, 100);
+        double actual = craps.hardWays(5, 5, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -283,10 +283,10 @@ public class CrapsTest {
     public void testHardWays4(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.hardWays(2, 2, 100);
+        double actual = craps.hardWays(2, 2, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -296,10 +296,10 @@ public class CrapsTest {
     public void testHardWaysLoss(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 0;
+        double expected = 0;
 
         //When
-        int actual = craps.hardWays(1, 1, 100);
+        double actual = craps.hardWays(1, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -309,10 +309,10 @@ public class CrapsTest {
     public void testOneRollBets1(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 3000;
+        double expected = 3000;
 
         //When
-        int actual = craps.oneRollBets(1, 1, 100);
+        double actual = craps.oneRollBets(1, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -322,10 +322,10 @@ public class CrapsTest {
     public void testOneRollBets2(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 3000;
+        double expected = 3000;
 
         //When
-        int actual = craps.oneRollBets(6, 6, 100);
+        double actual = craps.oneRollBets(6, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -335,10 +335,10 @@ public class CrapsTest {
     public void testOneRollBets3(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 1500;
+        double expected = 1500;
 
         //When
-        int actual = craps.oneRollBets(1, 2, 100);
+        double actual = craps.oneRollBets(1, 2, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -348,10 +348,10 @@ public class CrapsTest {
     public void testOneRollBets4(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 1500;
+        double expected = 1500;
 
         //When
-        int actual = craps.oneRollBets(2, 1, 100);
+        double actual = craps.oneRollBets(2, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -361,10 +361,10 @@ public class CrapsTest {
     public void testOneRollBets5(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 1500;
+        double expected = 1500;
 
         //When
-        int actual = craps.oneRollBets(5, 6, 100);
+        double actual = craps.oneRollBets(5, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -374,10 +374,10 @@ public class CrapsTest {
     public void testOneRollBets6(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 1500;
+        double expected = 1500;
 
         //When
-        int actual = craps.oneRollBets(6, 5, 100);
+        double actual = craps.oneRollBets(6, 5, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -387,10 +387,10 @@ public class CrapsTest {
     public void testOneRollBetsLoss(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 0;
+        double expected = 0;
 
         //When
-        int actual = craps.oneRollBets(4, 6, 100);
+        double actual = craps.oneRollBets(4, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -400,10 +400,10 @@ public class CrapsTest {
     public void testAnyCrapsBet1(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anyCrapsBet(1, 1, 100);
+        double actual = craps.anyCrapsBet(1, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -413,10 +413,10 @@ public class CrapsTest {
     public void testAnyCrapsBet2(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anyCrapsBet(1, 2, 100);
+        double actual = craps.anyCrapsBet(1, 2, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -426,10 +426,10 @@ public class CrapsTest {
     public void testAnyCrapsBet3(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anyCrapsBet(2, 1, 100);
+        double actual = craps.anyCrapsBet(2, 1, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -439,10 +439,10 @@ public class CrapsTest {
     public void testAnyCrapsBet4(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 700;
+        double expected = 700;
 
         //When
-        int actual = craps.anyCrapsBet(6, 6, 100);
+        double actual = craps.anyCrapsBet(6, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
@@ -452,10 +452,10 @@ public class CrapsTest {
     public void testAnyCrapsBetLoss(){
         //Given
         CrapsGame craps = new CrapsGame();
-        int expected = 0;
+        double expected = 0;
 
         //When
-        int actual = craps.anyCrapsBet(4, 6, 100);
+        double actual = craps.anyCrapsBet(4, 6, 100);
 
         //Then
         Assert.assertEquals(expected, actual);
